fix(arm32): cast math.MaxUint32 to int64 for 32-bit int comparison
On 32-bit ARM (GOARCH=arm), Go's int is 32-bit, so the bare comparison
`v > math.MaxUint32` (where MaxUint32 is the untyped constant 4294967295)
fails to compile with "math.MaxUint32 (untyped int constant 4294967295)
overflows int". Cast both sides to int64 to keep the bounds check semantically
correct without changing behavior on 64-bit platforms.

Affects:
- cycletls/packet_builder.go writeU32
- cycletls/server/protocol/encoder.go safeWriteUint32

Unblocks build:go:linux:arm in CI.

+++ cycletls/packet_builder.go
@@ -65,7 +65,7 @@ func buildEndFrame(requestID string) []byte {
 // writeU32 writes a 32-bit unsigned integer in big-endian format.
 // Returns an error if v is outside the uint32 range [0, 4294967295].
 func writeU32(b *bytes.Buffer, v int) error {
-	if v < 0 || v > math.MaxUint32 {
+	if v < 0 || int64(v) > int64(math.MaxUint32) {
 		return ErrU32Overflow
 	}
 	b.WriteByte(byte(v >> 24))
